leaf

+++ Shared/Utilities.c
@@ -25,7 +25,7 @@ float rr_angle_lerp(float start, float end, float t)
     {
         if (end > start)
             start += 2 * M_PI;
-        else 
+        else
             end += 2 * M_PI;
         return fmod((end - start) * t + start + 2 * M_PI, 2 * M_PI);
     }
@@ -33,5 +33,5 @@ float rr_angle_lerp(float start, float end, float t)
 
 float rr_frand()
 {
-    return (float) rand() / RAND_MAX;
+    return (double)rand() / (double)RAND_MAX;
 }
